refactor: decompose LeidenCommunityDetector (470→215+78+148) into partials
Split by responsibility: orchestration, graph construction, algorithm execution.
